Remove Statement#log_only_transactions
log_only_transactions?(log) was providing some over the top rigger to
what was allowed into Statement. a) It's not central to the user
stories so creating more code than requested. b) The knock on impact
of the method was that it was about to make all my tests very rigid,
so I've removed it. Though of I'm still mocking transactions
correctly in the tests.